Create a module to handle errors messages
