+++ acc.csv
@@ -1,0 +1,8 @@
+1,10000
+2,200
+3,20
+4,1
+5,
+
+6,90
+,8
