+++ main.rb
@@ -2,7 +2,7 @@ require_relative 'services/csv/open_files'
 
 class Main
   def initialize
-    @accounts_csv, @transaction_csv = Services::CSV::OpenFiles.new(ARGV).files
+    @accounts_csv, @transaction_csv = Services::CSV::OpenFiles.call(ARGV)
   end
 end
