+++ trans.csv
@@ -1,0 +1,4 @@
+1,-10
+2,1
+4,-5
+1,-10000
